evaluation with the nine runs
Each pickle file contains a dictionary where keys are queries and values are the ranked list of results. This file will generate spreadsheets 11 spreadhseets. 9 for precision and recall for each of the nine runs (52 queries each). 1 for the MAP and MRR values for all runs. 1 for the P@K values for all runs.